requirements.txt
aiofiles==25.1.0
aiogram==3.27.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-types==0.7.0
asgiref==3.11.1
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.6
Django==6.0.3
django-jazzmin @ git+https://github.com/farridav/django-jazzmin.git@8b5fa3726fa86084ea4e433378f1bbc2c1c1e7a5
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.15
frozenlist==1.8.0
idna==3.11
inflection==0.5.1
magic-filter==1.0.12
multidict==6.7.1
packaging==26.0
pillow==12.1.1
propcache==0.4.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.12.1
python-decouple==3.8
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.33.1
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
yarl==1.23.0
